+++ SavchenkoProj/src/main/java/org/pages/ParentPage.java
@@ -1,9 +1,11 @@
 package org.pages;
 
+import org.apache.log4j.Logger;
 import org.junit.Assert;
 import org.openqa.selenium.WebDriver;
 
 abstract public class ParentPage extends CommonActionsWithElements {
+    private Logger logger = Logger.getLogger(getClass());
     protected String baseUrl = "https://aqa-complexapp.onrender.com";
 
     public ParentPage(WebDriver webDriver) {
@@ -23,5 +25,48 @@ abstract public class ParentPage extends CommonActionsWithElements {
     protected void checkUrlWithPattern() {
         Assert.assertTrue("URL is not expected" + "Expected URL: " + baseUrl + getRelativeUrl() + "Actual URL: " + webDriver.getCurrentUrl(),
                 webDriver.getCurrentUrl().matches(baseUrl + getRelativeUrl()));
+    }
+
+    private void printErrorAndStopTest(Exception e) {
+        logger.error("Cannot work with element " + e);
+        Assert.fail("Cannot work with element " + e);
+    }
+
+    public void switchToNewTab() {
+        try {
+            for (String winHandle : webDriver.getWindowHandles()) {
+                webDriver.switchTo().window(winHandle);
+            }
+            logger.info("Switched to new tab");
+        } catch (Exception e) {
+            printErrorAndStopTest(e);
+        }
+    }
+
+    public void switchToFirstTab() {
+        try {
+            webDriver.switchTo().defaultContent();
+            logger.info("Switched to first tab");
+        } catch (Exception e) {
+            printErrorAndStopTest(e);
+        }
+    }
+
+    public void closeCurrentTab() {
+        try {
+            webDriver.close();
+            logger.info("New tab was closed");
+        } catch (Exception e) {
+            printErrorAndStopTest(e);
+        }
+    }
+
+    public void refreshPage() {
+        try {
+            webDriver.navigate().refresh();
+            logger.info("Page was refreshed");
+        } catch (Exception e) {
+            printErrorAndStopTest(e);
+        }
     }
 }
